+++ src/app/loadout/random/random-loadout.component.js
@@ -1,5 +1,5 @@
-import angular from 'angular';
 import _ from 'underscore';
+import { optimalLoadout } from '../loadout-utils';
 import template from './random-loadout.html';
 import './random-loadout.scss';
 
@@ -11,7 +11,7 @@ export const RandomLoadoutComponent = {
   }
 };
 
-function RandomLoadoutCtrl($window, $scope, dimStoreService, D2StoresService, dimLoadoutService, $i18next) {
+function RandomLoadoutCtrl($window, dimStoreService, D2StoresService, dimLoadoutService, $i18next) {
   'ngInject';
 
   const vm = this;
@@ -26,35 +26,20 @@ function RandomLoadoutCtrl($window, $scope, dimStoreService, D2StoresService, di
   vm.applyRandomLoadout = function(e) {
     e.preventDefault();
 
-    if (vm.disableRandomLoadout) {
+    if (vm.disableRandomLoadout || !$window.confirm($i18next.t('Loadouts.Randomize'))) {
       return null;
     }
-
-    if (!$window.confirm($i18next.t('Loadouts.Randomize'))) {
-      return null;
-    }
-
-    vm.disableRandomLoadout = true;
 
     const store = _.find(vm.stores, 'current');
     if (!store) {
       return null;
     }
 
-    const types = store.destinyVersion === 1 ? [
+    const types = new Set([
       'Class',
       'Primary',
       'Special',
       'Heavy',
-      'Helmet',
-      'Gauntlets',
-      'Chest',
-      'Leg',
-      'ClassItem',
-      'Artifact',
-      'Ghost'
-    ] : [
-      'Class',
       'Kinetic',
       'Energy',
       'Power',
@@ -63,37 +48,20 @@ function RandomLoadoutCtrl($window, $scope, dimStoreService, D2StoresService, di
       'Chest',
       'Leg',
       'ClassItem',
-      'Ghost',
-    ];
+      'Artifact',
+      'Ghost'
+    ]);
+    const storeService = (store.destinyVersion === 2 ? D2StoresService : dimStoreService);
 
-    const accountItems = (store.destinyVersion === 2 ? D2StoresService : dimStoreService).getAllItems();
-    const items = {};
+    // Any item equippable by this character in the given types
+    const applicableItems = storeService.getAllItems().filter((i) => types.has(i.type) && i.canBeEquippedBy(store));
 
-    const foundExotic = {};
+    // Use "random" as the value function
+    const loadout = optimalLoadout(applicableItems, () => Math.random(), $i18next.t('Loadouts.Random'));
 
-    const fn = (type) => (item) => (item.type === type &&
-                                    item.canBeEquippedBy(store) &&
-                                    (item.typeName !== 'Mask' || ((item.typeName === 'Mask') && (item.tier === 'Legendary'))) &&
-                                    (!foundExotic[item.bucket.sort] || (foundExotic[item.bucket.sort] && !item.isExotic)));
-
-    _.each(types, (type) => {
-      const filteredItems = _.filter(accountItems, fn(type));
-      const random = filteredItems[Math.floor(Math.random() * filteredItems.length)];
-
-      if (!foundExotic[random.bucket.sort]) {
-        foundExotic[random.bucket.sort] = random.isExotic;
-      }
-
-      const clone = angular.extend(angular.copy(random), { equipped: true });
-      items[type.toLowerCase()] = [clone];
-    });
-
+    vm.disableRandomLoadout = true;
     return dimLoadoutService
-      .applyLoadout(store, {
-        classType: -1,
-        name: $i18next.t('Loadouts.Random'),
-        items: items
-      }, true)
+      .applyLoadout(store, loadout, true)
       .finally(() => {
         vm.disableRandomLoadout = false;
       });